refactor(tangle-dapp): Cleanup locked balance details table
