+++ apps/tangle-dapp/data/democracy/useDemocracy.ts
@@ -1,4 +1,4 @@
-import { useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
 import { map } from 'rxjs';
 
 import { SubstrateLockId } from '../../constants';
@@ -6,8 +6,12 @@ import usePolkadotApiRx from '../../hooks/usePolkadotApiRx';
 import useBalancesLock from '../balances/useBalancesLock';
 
 const useDemocracy = () => {
-  const { data: votes } = usePolkadotApiRx((api, activeSubstrateAddress) =>
-    api.query.democracy.votingOf(activeSubstrateAddress)
+  const { data: votes } = usePolkadotApiRx(
+    useCallback(
+      (api, activeSubstrateAddress) =>
+        api.query.democracy.votingOf(activeSubstrateAddress),
+      []
+    )
   );
 
   const latestDirectVote = useMemo(() => {
@@ -20,31 +24,38 @@ const useDemocracy = () => {
     return votes.asDirect.votes[votes.asDirect.votes.length - 1];
   }, [votes]);
 
-  const latestReferendumIndex = useMemo(() => {
+  const latestReferendumIndex = (() => {
     if (latestDirectVote === null) {
       return null;
     }
 
     return latestDirectVote[0];
-  }, [latestDirectVote]);
+  })();
 
-  const { data: latestReferendum } = usePolkadotApiRx((api) => {
-    if (latestReferendumIndex === null) {
-      return null;
-    }
+  const { data: latestReferendum } = usePolkadotApiRx(
+    useCallback(
+      (api) => {
+        if (latestReferendumIndex === null) {
+          return null;
+        }
 
-    return (
-      api.query.democracy
-        .referendumInfoOf(latestReferendumIndex)
-        // The referendum information should be defined if the latest
-        // referendum index is defined.
-        .pipe(map((referendumOpt) => referendumOpt.unwrap()))
-    );
-  });
+        return (
+          api.query.democracy
+            .referendumInfoOf(latestReferendumIndex)
+            // The referendum information should be defined if the latest
+            // referendum index is defined.
+            .pipe(map((referendumOpt) => referendumOpt.unwrap()))
+        );
+      },
+      [latestReferendumIndex]
+    )
+  );
 
   const { amount: lockedBalance } = useBalancesLock(SubstrateLockId.DEMOCRACY);
+  const isInDemocracy = lockedBalance === null ? null : lockedBalance.gtn(0);
 
   return {
+    isInDemocracy,
     lockedBalance,
     latestReferendum,
   };
