+++ apps/tangle-dapp/utils/index.ts
@@ -1,7 +1,9 @@
 export * from './calculateInflation';
-export * from './convertToSubstrateAddress';
+export * from './evmToSubstrateAddress';
+export * from './formatBnWithCommas';
 export { default as getChartDataAreaColorByServiceType } from './getChartDataAreaColorByServiceType';
 export { default as getChipColorOfServiceType } from './getChipColorOfServiceType';
 export { default as getMonthlyHistorySinceDate } from './getMonthlyHistorySinceDate';
 export { default as getRoundedDownNumberWith2Decimals } from './getRoundedDownNumberWith2Decimals';
 export * from './splitTokenValueAndSymbol';
+export * from './substrateToEvmAddress';
